remove class job manager server
